added some changes on athlete and started to calculate objective measurements

+++ src/main/java/com/adoo/colemanfitness/model/entity/BodyMeasurement.java
@@ -19,7 +19,7 @@ public class BodyMeasurement {
     private float weight;
     private float muscleMass;
     private float bodyFat;
-    private float height;
+    private Long height;
     private Date date;
     @ManyToOne
     @JoinColumn(name = "athlete_id")
